+++ packages/pigeon/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 9.0.7
 
 * [swift] Changes all ints to int64.
+  May require code updates to existing code.
 * Adds integration tests for int64.
 
 ## 9.0.6